chore: migrate create_reward_center into create/handler.rs
